Helper classes moved to new file
