+++ plugins/csharp/parser/src_csharp/Program.cs
@@ -195,101 +195,6 @@ namespace CSharpParser
                 return index;
             });
             return await ParsingTask;
-        }
-
-        
-    }
-
-    class ProgramHelper
-    {
-        public static IEnumerable<string> GetSourceFilesFromDir(string root, string extension)
-        {
-            IEnumerable<string> allFiles = new string[]{};
-            // Data structure to hold names of subfolders. 
-            ArrayList dirs = new ArrayList();
-
-            if (!System.IO.Directory.Exists(root))
-            {
-                throw new ArgumentException();
-            }
-            dirs.Add(root);
-
-            while (dirs.Count > 0)
-            {
-                string currentDir = dirs[0].ToString();
-                dirs.RemoveAt(0);
-                string[] subDirs;
-                try
-                {
-                    subDirs = System.IO.Directory.GetDirectories(currentDir);
-                }
-                catch (UnauthorizedAccessException e)
-                {
-                    WriteLine(e.Message);
-                    continue;
-                }
-                catch (System.IO.DirectoryNotFoundException e)
-                {
-                    WriteLine(e.Message);
-                    continue;
-                }
-
-                // Add the subdirectories for traversal.
-                dirs.AddRange(subDirs);
-
-                string[] files = null;
-                try
-                {
-                    files = System.IO.Directory.GetFiles(currentDir);
-                }
-                catch (UnauthorizedAccessException e)
-                {
-                    Console.WriteLine(e.Message);
-                    continue;
-                }
-                catch (System.IO.DirectoryNotFoundException e)
-                {
-                    Console.WriteLine(e.Message);
-                    continue;
-                }
-
-                foreach (string file in files)
-                {
-                    try
-                    {
-                        System.IO.FileInfo fi = new System.IO.FileInfo(file);
-                        if (fi.Extension == extension) {
-                            allFiles = allFiles.Append(file);
-                        }
-                    }
-                    catch (System.IO.FileNotFoundException e)
-                    {
-                        // If file was deleted by a separate application
-                        Console.WriteLine(e.Message);
-                    }
-                }
-            }
-
-            return allFiles;
-        }
-
-        public static string transformConnectionString(string _connectionString)
-        {
-            _connectionString = _connectionString.Substring(_connectionString.IndexOf(':')+1);
-            _connectionString = _connectionString.Replace("user", "username");
-            string [] properties = _connectionString.Split(';');
-            string csharpConnectionString = "";
-            for (int i = 0; i < properties.Length; ++i)
-            {
-                csharpConnectionString += properties[i].Substring(0,1).ToUpper()
-                    + properties[i].Substring(1);
-                if (i < properties.Length-1)
-                {
-                    csharpConnectionString += ";";
-                }
-            }
-
-            return csharpConnectionString;
-        }
+        }   
     }
 }
